feat(inventory): support fractional quantities and unit merging
  quantity was Int32 across model/store/tools/parser, so "500 gram"
  truncated and couldn't be combined with "5 kg". Switch to Double
  everywhere and merge on add: same item name + compatible unit
  (kg/gram, liter/ml) sums into one entry via a shared base unit
  instead of creating duplicate rows.

  Also switch inventory list from date-grouped sections to a flat
  alphabetical list, now that merged entries make date grouping less
  meaningful for "how much do we have" browsing.

+++ stock-inventory-ai-ios/Services/LLMService.swift
@@ -202,7 +202,7 @@ final class LLMService {
 
     struct ParsedStockEntry: Decodable {
         let itemName: String
-        let quantity: Int
+        let quantity: Double
         let unit: String
     }
